refactor: repository 테스트용 db 변경 - 테스트에 h2 db 사용하면서 profile 추가 - 어노테이션 변경 - repository 인터페이스 없애고 통일
