+++ src/main/resources/application.properties
@@ -2,4 +2,3 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.url=jdbc:mysql://localhost:3306/chess_db?serverTimezone=UTC&characterEncoding=UTF-8
 spring.datasource.username=root
 spring.datasource.password=root
-spring.profiles.active=prod
