+++ src/test/java/chess/repository/ChessRepositoryTest.java
@@ -11,19 +11,22 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
 
-@SpringBootTest()
+@JdbcTest
 @ActiveProfiles("test")
 public class ChessRepositoryTest {
 
-    @Autowired
     private ChessRepository chessRepository;
+
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
 
     @BeforeEach
     void setUp() {
-        chessRepository.deleteAll();
+        chessRepository = new ChessRepository(jdbcTemplate);
     }
 
     @Test
