+++ src/test/resources/application-test.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:h2:tcp://localhost/~/test;MODE=MySQL;DB_CLOSE_DELAY=-1
+spring.datasource.url=jdbc:h2:tcp://localhost/~/chess_db;MODE=MySQL;DB_CLOSE_DELAY=-1
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=
